feat: make SimpleArena work with Allocator + complex objects
When I was designing ObjectArena, I thought the Arena was the one in
charge of initializing and destroying the objects stored in a container.
But it turns out that, as explained in the Allocator reference in
cppreference.com, it is the container itself the one in charge of
doing so.

With this commit, I prove exactly that: I removed the compilation
restriction that limited SimpleArena to only trivially destructible
types and test that it indeed works with a user-defined type.